requirements.txt.txt
requests
jinja2
python-dotenv
pyfiglet
colorama